Extract cloud backup components
